* Move `Output` classes to a separate file. * `config` also refactored to be inside `compiler`. * constants moved into their own file. * warnings context managers moved into the `utils` file
