+++ labscript/constants.py
@@ -1,0 +1,8 @@
+ns = 1e-9
+us = 1e-6
+ms = 1e-3
+s = 1
+Hz = 1
+kHz = 1e3
+MHz = 1e6
+GHz = 1e9
